Refactor generic app_utils to util module
