+++ elyra/util/__init__.py
@@ -15,4 +15,6 @@
 #
 
 from .archive import create_temp_archive
+from .cos import CosClient
 from .path import get_absolute_path, get_expanded_path
+from .app_utils import CliOption, Flag, Option
